requirements.txt
python-telegram-bot==20.8
google-generativeai
gspread
oauth2client
google-api-python-client
nest_asyncio
